requirements.txt
flask
gunicorn
flask-cors
python-dotenv
boto3